Moved models serialization from MessageExtension to its methods

+++ src/Client/entities/LivingEntity.cs
@@ -2,6 +2,6 @@ namespace ValorRise.Client.Entities;
 
 public abstract class LivingEntity : Entity
 {
-    public double Health { get; set; }
-    public double MaxHealth { get; set; }
+    public float Health { get; set; }
+    public float MaxHealth { get; set; }
 }
